feat: 서비스 추가 개발

+++ server/src/refund/refund.service.ts
@@ -12,31 +12,46 @@ export class RefundService {
    * @param createRefundDto
    * @return boolean 값으로 리턴 'true'/'false'
    */
-  async create(createRefundDto: CreateRefundDto): Promise<boolean> {
+  async create(
+    order_productCode: string,
+    createRefundDto: CreateRefundDto
+  ): Promise<boolean> {
+    const [con, code] = await Promise.all([
+      this.databaseService.getConnection(),
+      this.databaseService.genCode(),
+    ]);
+    if (!con) {
+      throw new Error();
+    }
     try {
-      const [con, code] = await Promise.all([
-        this.databaseService.getConnection(),
-        this.databaseService.genCode(),
-      ]);
+      await con.beginTransaction();
       await con.query(`
           INSERT INTO refunds
           (code,
             amount,
-            name,
             reason_type,
             reason_detail,
             fk_order_code,
-            fk_site_code)
+            fk_site_code
+            )
           VALUES ('${code}',
           '${createRefundDto.amount}',
           '${createRefundDto.reason_type}',
           '${createRefundDto.reason_detail}',
-          '${createRefundDto.fk_order_code}'
+          '${createRefundDto.fk_order_code}',
+          '${createRefundDto.fk_site_code}')
           `);
-      con.release();
+      const refundOrderProductCode = await this.databaseService.genCode();
+      await con.query(`
+          INSERT INTO refund_order_product(code, fk_order_product_code, fk_refund_code) VALUES('${refundOrderProductCode}', '${order_productCode}','${code}')
+      `);
+      await con.commit();
       return true;
     } catch (error) {
+      con.release();
       throw error;
+    } finally {
+      con.release();
     }
   }
 
@@ -54,6 +69,9 @@ export class RefundService {
   ): Promise<GetRefundDto[]> {
     console.log(code);
     const con = await this.databaseService.getConnection();
+    if (!con) {
+      throw new Error();
+    }
     try {
       const [row] = await con.query(`
     SELECT idx FROM refunds WHERE code = '${code}' &&fk_site_code ='${site_code}'
@@ -74,13 +92,85 @@ export class RefundService {
         refunds.push(refundData[i]);
       }
       console.log(refunds);
-      con.release();
+
       return refunds;
     } catch (error) {
+      con.release();
       throw error;
+    } finally {
+      con.release();
+    }
+  }
+  /**Order 정보 조회
+   */
+  async findOrderInfo(code: string) {
+    const con = await this.databaseService.getConnection();
+    try {
+      const [refundData] = await con.query(`
+      SELECT O.code,
+      O.order_no,
+      O.site_code,
+      O.user_code,
+      O.post_number,
+      O.receiver_name,
+      O.receiver_address,
+      O.receiver_phone,
+      O.receiver_phone2,
+      O.status,
+      O.total_price
+        from refunds R
+        inner join refund_order_product ROP
+        on R.code = ROP.fk_refund_code
+        inner join order_product OP
+        on ROP.fk_order_product_code = OP.code
+        inner join orders O
+        on OP.fk_order_code = O.code
+      WHERE R.code = '${code}'
+      `);
+      console.log(refundData);
+      return refundData[0];
+    } catch (error) {
+      con.release();
+      throw error;
+    } finally {
+      con.release();
     }
   }
 
+  /**Product 정보 조회 */
+  async findProductInfo(code: string) {
+    const con = await this.databaseService.getConnection();
+    try {
+      const [refundData] = await con.query(`
+      SELECT P.code,
+      P.price,
+      P.updated_time,
+      P.name,
+      P.prodStatus,
+      P.created_time,
+      P.stock,
+      P.image_url,
+      P.description,
+      P.category,
+      P.fk_site_code
+        from refunds R
+        inner join refund_order_product ROP
+        on R.code = ROP.fk_refund_code
+        inner join order_product OP
+        on ROP.fk_order_product_code = OP.code
+        inner join products P
+        on OP.fk_product_code = P.code
+      WHERE R.code = '${code}'
+      `);
+      console.log(refundData);
+      return refundData[0];
+    } catch (error) {
+      con.release();
+      throw error;
+    } finally {
+      con.release();
+    }
+  }
   /**
    * code값으로 환불 단일 조회
    * @param code
@@ -104,33 +194,49 @@ export class RefundService {
    * @param updateRefundDto
    */
   async update(code: string, updateRefundDto: UpdateRefundDto) {
+    const con = await this.databaseService.getConnection();
+    if (!con) {
+      throw new Error();
+    }
     try {
-      const refundData = await this.databaseService.query(`
-      SELECT * FROM refunds WHERE code = '${code}'`);
+      console.log(code);
+      const [refundData] = await con.query(`
+      SELECT * FROM refunds WHERE code='${code}'
+      `);
+      console.log(refundData);
       const refund: GetRefundDto = refundData[0];
-      const newRefund = await this.databaseService.query(`
-    UPDATE refunds
+      const amount = updateRefundDto.amount
+        ? updateRefundDto.amount
+        : refund.amount;
+      const reason_type = updateRefundDto.reason_type
+        ? updateRefundDto.reason_type
+        : refund.reason_type;
+      const reason_detail = updateRefundDto.reason_detail
+        ? updateRefundDto.reason_detail
+        : refund.reason_detail;
+      const fk_order_code = updateRefundDto.fk_order_code
+        ? updateRefundDto.fk_order_code
+        : refund.fk_order_code;
+      const fk_site_code = updateRefundDto.fk_site_code
+        ? updateRefundDto.fk_site_code
+        : refund.fk_site_code;
+
+      await con.query(`
+    UPDATE refunds 
     SET 
-    price =  IF(${updateRefundDto.amount != undefined},'${
-        updateRefundDto.amount
-      }','${refund.amount}'),
-    name =  IF(${updateRefundDto.reason_type != undefined},'${
-        updateRefundDto.reason_type
-      }','${refund.reason_type}'),
-    prodStatus = IF(${updateRefundDto.reason_detail != undefined},'${
-        updateRefundDto.reason_detail
-      }','${refund.reason_detail}'),
-    stock = IF(${updateRefundDto.fk_order_code != undefined},'${
-        updateRefundDto.fk_order_code
-      }','${refund.fk_order_code}'),
-    image_url = IF(${updateRefundDto.fk_site_code != undefined},'${
-        updateRefundDto.fk_site_code
-      }','${refund.fk_site_code}')
+    amount = '${amount}',
+    reason_type = '${reason_type}',
+    reason_detail = '${reason_detail}',
+    fk_order_code='${fk_order_code}',
+    fk_site_code='${fk_site_code}'
     WHERE code='${code}';
     `);
-      return newRefund[0];
+      return true;
     } catch (error) {
+      con.release();
       throw error;
+    } finally {
+      con.release();
     }
   }
 
@@ -139,12 +245,19 @@ export class RefundService {
    * @param code
    */
   async remove(code: string) {
+    const con = await this.databaseService.getConnection();
+    if (!con) {
+      throw new Error();
+    }
     try {
-      const refund = await this.databaseService.query(`
+      await con.query(`
     DELETE FROM refunds WHERE code='${code}';`);
-      return refund[0];
+      return true;
     } catch (error) {
+      con.release();
       throw error;
+    } finally {
+      con.release();
     }
   }
 }
